test/jian_digitalhuman_mcp-dev/requirements.txt
fastmcp>=2.6.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
cryptography>=41.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0 